Remove scale_samples; store samples as float
The heuristic that separates between time/depth is broken, since the
sample rate for depth is millimeters, not meters. In other words, it is
impossible to figure out just by inspecting the file if sample interval
means time or depth. The entire tool is removed in favour of correctly
storing the samples as floats.

+++ python/segyio/create.py
@@ -69,7 +69,7 @@ def create(filename, spec):
     """
     f = segyio.SegyFile(filename, "w+")
 
-    f._samples       = numpy.asarray(spec.samples, dtype = numpy.intc)
+    f._samples       = numpy.asarray(spec.samples, dtype = numpy.single)
     f._ext_headers   = spec.ext_headers
     f._bsz           = _segyio.trace_bsize(len(f.samples))
